make stateboard completely requireable
