+++ CMakeLists.txt
@@ -165,18 +165,6 @@ install(
   DESTINATION ${TARANTOOL_INSTALL_LUADIR}
 )
 
-# Make stateboard.lua executable and requireable at the same time
-install(
-  FILES ${CMAKE_CURRENT_SOURCE_DIR}/stateboard.lua
-  DESTINATION ${TARANTOOL_INSTALL_LUADIR}
-)
-
-install(
-  FILES ${CMAKE_CURRENT_SOURCE_DIR}/stateboard.lua
-  DESTINATION ${TARANTOOL_INSTALL_BINDIR}
-  RENAME stateboard
-)
-
 install(
   FILES
     ${CMAKE_CURRENT_BINARY_DIR}/VERSION.lua
