+++ test/entrypoint/srv_stateboard.lua
@@ -1,0 +1,5 @@
+#!/usr/bin/env tarantool
+
+require('strict').on()
+local stateboard = require('cartridge.stateboard')
+stateboard.start()
